Fix issue with the HTML5 color widget defaulting to black for named colors
The HTML5 color input only works with long hexadecimal color notation. When converting a Color object to string, the colour package will automatically convert to the smallest possible web format including named colors and short hex notation, resulting in the default black color in the color widget. This change forces the value to be written in long hex format.